Enable scratchpad for GRU (#2366)
* Enable scratchpad for GRU
* Change scratchpaad size from int to size_t

---------

+++ csrc/gpu/oneDNN/BatchNorm.h
@@ -178,7 +178,7 @@ static std::tuple<at::Tensor&, at::Tensor&, at::Tensor&> batch_normalization(
   args.insert({DNNL_ARG_VARIANCE, var_m});
 
 #ifdef USE_SCRATCHPAD_MODE
-  int scratchpad_size = bn_fwd_pd.scratchpad_desc().get_size();
+  size_t scratchpad_size = bn_fwd_pd.scratchpad_desc().get_size();
   Tensor scratchpad_tensor = at::AtenIpexTypeXPU::empty(
       {scratchpad_size}, src.options().dtype(at::kByte), c10::nullopt);
   auto scratchpad_memory = dpcpp_onednn_memory(
@@ -314,7 +314,7 @@ batch_normalization_backward(
   };
 
 #ifdef USE_SCRATCHPAD_MODE
-  int scratchpad_size = bn_bwd_pd.scratchpad_desc().get_size();
+  size_t scratchpad_size = bn_bwd_pd.scratchpad_desc().get_size();
   Tensor scratchpad = at::AtenIpexTypeXPU::empty(
       {scratchpad_size}, src.options().dtype(at::kByte), c10::nullopt);
   auto scratchpad_m = dpcpp_onednn_memory(
